requirements.txt
dotenv
groq
mistralai
request
PIL
